+++ composer.json
@@ -1,7 +1,7 @@
 {
     "name": "macocci7/php-scatterplot",
     "description": "it's easy to use for creating scatter plots.",
-    "version": "1.2.0",
+    "version": "1.2.1",
     "type": "library",
     "license": "MIT",
     "autoload": {
@@ -18,7 +18,7 @@
     "minimum-stability": "stable",
     "require": {
         "php": ">=8.1",
-        "intervention/image": "^3.5",
+        "intervention/image": "^3.6",
         "macocci7/php-frequency-table": "^1.3",
         "macocci7/php-csv": "^1.0",
         "macocci7/php-combination": "^1.0",
@@ -29,6 +29,6 @@
         "phpmd/phpmd": "^2.15",
         "phpunit/phpunit": "^10.5",
         "phpstan/phpstan": "^1.10",
-        "php-parallel-lint/php-parallel-lint": "^1.3"
+        "php-parallel-lint/php-parallel-lint": "^1.4"
     }
 }
