+++ methods.txt
@@ -1,0 +1,54 @@
+Core:
+     1	    public function __construct()
+     2	    private function loadConf()
+     3	    public function config(string|array $configResource)
+     4	    public function layer(array $layer)
+     5	    public function layers(array $layers)
+    29	    public function getConfig(string|null $key = null)
+
+Attribute:
+     6	    public function limitX(int|float $lower, int|float $upper)
+     7	    public function limitY(int|float $lower, int|float $upper)
+     8	    public function resize(int $width, int $height)
+     9	    public function frame($xRatio, $yRatio)
+    25	    public function labelX(string $label)
+    26	    public function labelY(string $label)
+    27	    public function caption(string $caption)
+    28	    public function legends(array $legends)
+
+StyleCore:
+    10	    public function bgcolor(string|null $color = null)
+    11	    public function axis(int $width, string|null $color = null)
+    12	    public function grid(int $width, string|null $color = null)
+    13	    public function gridXPitch(int|float $pitch)
+    14	    public function gridYPitch(int|float $pitch)
+
+    15	    public function colors(array $colors)
+    16	    public function plotSize(int $size)
+    17	    public function fontPath(string $path)
+    18	    public function fontSize(int|float $size)
+    19	    public function fontColor(string $color)
+
+StyleAppendix:
+    20	    public function referenceLineX(int|float $x, int $width = 1, string $color = '#0000ff')
+    21	    public function referenceLineY(int|float $y, int $width = 1, string $color = '#0000ff')
+    22	    public function specificationLimitX(int|float $lower, int|float $upper, int $width = 1, string $color = '#ff00ff')
+    23	    public function specificationLimitY(int|float $lower, int|float $upper, int $width = 1, string $color = '#ff00ff')
+    24	    public function regressionLine(int $width, array $colors)
+
+VisibilityCore:
+    30	    public function gridXOn()
+    31	    public function gridXOff()
+    32	    public function gridYOn()
+    33	    public function gridYOff()
+    42	    public function legendOff()
+
+VisibilityAppendix:
+    34	    public function referenceLineXOff()
+    35	    public function referenceLineYOff()
+    36	    public function referenceLinesOff()
+    37	    public function specificationLimitXOff()
+    38	    public function specificationLimitYOff()
+    39	    public function specificationLimitsOff()
+    40	    public function regressionLineOn()
+    41	    public function regressionLineOff()
